async_api/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.1
orjson==3.4.1
isort==5.10.1
pydantic==1.9.0
uvicorn[standard]==0.17.6
gunicorn==20.1.0
uvloop==0.14.0
python-environ==0.4.54
